interface/requirements.txt
pandas
confluent-kafka
streamlit==1.45.0
matplotlib
psycopg2-binary
numpy
